Large code update
New post, two new grass textures, and tons of code - I just copied the
whole project from eclipse so it should be ready to run - it now
includes a first model of the branches and a working Main class that
generates 3 of them - read all about it in my latest post!
Also, a bit of documentation and some javadoc so that you will get a
better sense of what's going on.
A more full documentation may be updated in a couple of hours - so stay
tuned!
134 files to be commited, wow.