Full-scale trees
As I wrote yesterday in my post, leaves are now working and the tree
looks great! Try and run app.Main to see how it looks yourself.
Also in this update: Some old examples of apps (most of them aren't
currently working because of changes to the data structure) saved to
test.app.